Correctly copy wx on ubuntu and remove copy of netcdf.h into include
Python works as a framework on macos which means there are multiple include/python2.7 folders on a given machine (depending on how many version of sdk exist) and these are not directly in a python folder. Therefore, I am unsure how to make this work, so for now this will be excluded.